Simplify predict_genome after /simplify review (#118)
- Hoist GenomeWindowDataset to a shared module so the classifier and
  segmentation inference scripts stop carrying byte-for-byte copies.
  Cast chroms to str once in __init__ instead of per __getitem__ call.
- Gate torch.compile on torch.cuda.is_available() — avoids paying
  ~30s Inductor compile cost on CPU/test paths.
- Replace 'assert predictions is not None' with an explicit RuntimeError
  so it survives Python -O.
- Set TORCHINDUCTOR_CACHE_DIR to a repo-local dir in the prediction
  rule so subsequent genomes in a batch reuse the compile cache
  (~30s per genome * 20 genomes = ~10 min saved per full run).
